+++ libs/eavmlib/src/network.erl
@@ -27,7 +27,8 @@
     wait_for_ap/0, wait_for_ap/1, wait_for_ap/2,
     sta_rssi/0,
     sta_disconnect/0,
-    sta_connect/0, sta_connect/1
+    sta_connect/0, sta_connect/1,
+    sta_status/0
 ]).
 -export([start/1, start_link/1, stop/0]).
 -export([
@@ -163,12 +164,15 @@
 -type network_config() :: [sta_config() | ap_config() | sntp_config()].
 
 -type db() :: integer().
+-type sta_status() ::
+    associated | connected | connecting | degraded | disconnected | disconnecting | undefined.
 
 -record(state, {
     config :: network_config(),
     port :: port(),
     ref :: reference(),
-    sta_ip_info :: ip_info()
+    sta_ip_info :: ip_info(),
+    sta_state :: sta_status()
 }).
 
 %%-----------------------------------------------------------------------------
@@ -396,6 +400,24 @@ sta_rssi() ->
         Other -> {error, Other}
     end.
 
+%%-----------------------------------------------------------------------------
+%% @returns ConnectionState :: sta_status().
+%%
+%% @doc Get the connection status of the sta interface.
+%%
+%% Results will be one of: `associated', `connected', `connecting', `degraded',
+%% `disconnected', `disconnecting', or `undefined'. The state `associated' indicates
+%% that the station is connected to an access point, but does not yet have an IP address.
+%% A status of `degraded' indicates that the connection has experienced at least one
+%% beacon timeout event during the current connection session. This does not necessarily
+%% mean the connection is still in a poor state, but it might be helpful diagnosing
+%% problems with networked applications.
+%% @end
+%%-----------------------------------------------------------------------------
+-spec sta_status() -> Status :: sta_status().
+sta_status() ->
+    gen_server:call(?SERVER, sta_status).
+
 %%
 %% gen_server callbacks
 %%
@@ -409,17 +431,29 @@ handle_call(start, From, #state{config = Config} = State) ->
     Port = get_port(),
     Ref = make_ref(),
     Port ! {self(), Ref, {start, Config}},
-    wait_start_reply(Ref, From, Port, State);
+    case proplists:get_value(sta, Config) of
+        undefined ->
+            wait_start_reply(Ref, From, Port, State);
+        STA ->
+            case proplists:get_value(managed, STA) of
+                false ->
+                    wait_start_reply(Ref, From, Port, State#state{sta_state = connecting});
+                true ->
+                    wait_start_reply(Ref, From, Port, State#state{sta_state = disconnected})
+            end
+    end;
 handle_call(halt_sta, _From, #state{ref = Ref} = State) ->
     network_port ! {self(), Ref, halt_sta},
-    wait_halt_sta_reply(Ref, State);
+    wait_halt_sta_reply(Ref, State#state{sta_state = disconnecting});
 handle_call(connect, _From, #state{config = Config, ref = Ref} = State) ->
     network_port ! {self(), Ref, {connect, Config}},
-    wait_connect_reply(Ref, Config, State);
+    wait_connect_reply(Ref, Config, State#state{sta_state = connecting});
 handle_call({connect, Config}, _From, #state{config = OldConfig, ref = Ref} = State) ->
     NewConfig = update_config(OldConfig, Config),
     network_port ! {self(), Ref, {connect, NewConfig}},
-    wait_connect_reply(Ref, NewConfig, State);
+    wait_connect_reply(Ref, NewConfig, State#state{sta_state = connecting});
+handle_call(sta_status, _From, State) ->
+    {reply, State#state.sta_state, State};
 handle_call(_Msg, _From, State) ->
     {reply, {error, unknown_message}, State}.
 
@@ -431,22 +465,22 @@ wait_start_reply(Ref, From, Port, State) ->
             {noreply, State#state{port = Port, ref = Ref}};
         {Ref, {error, Reason} = ER} ->
             gen_server:reply(From, {error, Reason}),
-            {stop, {start_failed, Reason}, ER, State}
+            {stop, {start_failed, Reason}, ER, State#state{sta_state = disconnected}}
     end.
 
 %% @private
 wait_connect_reply(Ref, NewConfig, State) ->
     receive
         {Ref, ok} ->
-            {reply, ok, State#state{ref = Ref, config = NewConfig}};
+            {reply, ok, State#state{config = NewConfig}};
         {Ref, {error, _Reason} = ER} ->
-            {reply, ER, State}
+            {reply, ER, State#state{sta_state = disconnected}}
     end.
 
 wait_halt_sta_reply(Ref, State) ->
     receive
         {Ref, ok} ->
-            {reply, ok, State};
+            {reply, ok, State#state{sta_state = disconnected}};
         {Ref, {error, _Reason} = Error} ->
             {reply, Error, State}
     end.
@@ -458,16 +492,16 @@ handle_cast(_Msg, State) ->
 %% @hidden
 handle_info({Ref, sta_connected} = _Msg, #state{ref = Ref, config = Config} = State) ->
     maybe_sta_connected_callback(Config),
-    {noreply, State};
+    {noreply, State#state{sta_state = associated}};
 handle_info({Ref, sta_beacon_timeout} = _Msg, #state{ref = Ref, config = Config} = State) ->
     maybe_sta_beacon_timeout_callback(Config),
-    {noreply, State};
+    {noreply, State#state{sta_state = degraded}};
 handle_info({Ref, sta_disconnected} = _Msg, #state{ref = Ref, config = Config} = State) ->
     maybe_sta_disconnected_callback(Config),
-    {noreply, State};
+    {noreply, State#state{sta_state = disconnected, sta_ip_info = undefined}};
 handle_info({Ref, {sta_got_ip, IpInfo}} = _Msg, #state{ref = Ref, config = Config} = State) ->
     maybe_sta_got_ip_callback(Config, IpInfo),
-    {noreply, State#state{sta_ip_info = IpInfo}};
+    {noreply, State#state{sta_ip_info = IpInfo, sta_state = connected}};
 handle_info({Ref, ap_started} = _Msg, #state{ref = Ref, config = Config} = State) ->
     maybe_ap_started_callback(Config),
     {noreply, State};